# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD de Usuários",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Lista todos os usuários com paginação",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Lista de usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Cria um novo usuário",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Falha ao criar o usuário"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Retorna um usuário específico pelo ID",
                "operationId": "getUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/UserId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualiza um usuário",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "headers": {
                            "Accept": {
                                "description": "Tipo de resposta",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "headers": {
                            "Accept": {
                                "description": "Tipo de resposta",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Deleta um usuário específico pelo ID",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/UserId"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuário deletado com sucesso"
                    },
                    "400": {
                        "description": "Falha ao deletar o usuário"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateUserRequest": {
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "example": "joao@email.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "senha123"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "UpdateUserSchema": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Novu Nome"
                    },
                    "email": {
                        "type": "string",
                        "example": "novo@email.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "novaSenha123"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "example": "joao@email.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "senha123"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "UserId": {
                "name": "id",
                "in": "path",
                "description": "ID do usuário",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Usuários",
            "description": "Usuários"
        }
    ]
}